# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Anatomiya",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/term-join-image-file": {
            "get": {
                "tags": [
                    "Terms_Join_Image_File"
                ],
                "summary": "All term join image File",
                "description": "Hamma term join image filelarlarni ko'rish",
                "operationId": "term_join_image_file_index",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "term_id": {
                                            "type": "integer",
                                            "example": "5"
                                        },
                                        "image_file_id": {
                                            "type": "integer",
                                            "example": "4"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Terms_Join_Image_File"
                ],
                "summary": "Term_join_image File add",
                "description": "term join image filelarlar qo'shish",
                "operationId": "term_join_image_file_store",
                "requestBody": {
                    "description": "term join image file save",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "term_id",
                                    "image_file_id"
                                ],
                                "properties": {
                                    "term_id": {
                                        "type": "integer",
                                        "format": "number",
                                        "example": "6"
                                    },
                                    "image_file_id": {
                                        "type": "integer",
                                        "format": "number",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "term_id": {
                                            "type": "integer",
                                            "example": "Harley"
                                        },
                                        "image_file_id": {
                                            "type": "integer",
                                            "example": "Harley"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/term-join-image-file/{id}": {
            "get": {
                "tags": [
                    "Terms_Join_Image_File"
                ],
                "summary": "Show term join image file",
                "description": "bitta term join image fileda hamma malumotlarini ko'rsatadi",
                "operationId": "Terms_Join_Image_File_show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "term_join_image file id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "term_id": {
                                            "type": "integer",
                                            "example": "Harley"
                                        },
                                        "image_file_id": {
                                            "type": "integer",
                                            "example": "Harley"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Terms_Join_Image_File"
                ],
                "summary": "Update existing project",
                "description": "Returns updated project data",
                "operationId": "term_join_image_file_update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "term_join_image file id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "term update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "term_id",
                                    "image_file_id"
                                ],
                                "properties": {
                                    "term_id": {
                                        "type": "integer",
                                        "format": "number",
                                        "example": "6"
                                    },
                                    "image_file_id": {
                                        "type": "integer",
                                        "format": "number",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "term_id": {
                                            "type": "integer",
                                            "example": "Harley"
                                        },
                                        "image_file_id": {
                                            "type": "integer",
                                            "example": "Harley"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Terms_Join_Image_File"
                ],
                "summary": "Delete existing project",
                "description": "Deletes a record and returns no content",
                "operationId": "term_join_image_file_delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
    "security": {
        "bearer_token": {
            "type": "apiKey",
            "description": "Enter token in format (Bearer <token>)",
            "name": "Authorization",
            "in": "header"
        }
    }
}